Fixed env not loading on client side
