+++ src/app/layout.tsx
@@ -4,9 +4,9 @@ import LayoutManager from "./layouts"
 import { cookies } from "next/headers"
 import { cn } from "@/shared/lib/utils"
 import { Public_Sans } from "next/font/google"
-import { config } from "@/shared/utils/config"
+import { configUtil } from "@/shared/utils/config"
 
-config.validate()
+configUtil.validate()
 
 const publicSans = Public_Sans({
   variable: "--font-public-sans",
