+++ src/shared/schemas/env.ts
@@ -1,5 +1,5 @@
 import z from "zod"
 
 export const EnvSchema = z.object({
-  API_URL: z.url("Invalid API URL"),
+  NEXT_PUBLIC_API_URL: z.url("Invalid API URL"),
 })
