+++ src/shared/utils/config/type.ts
@@ -1,0 +1,6 @@
+import { EnvSchema } from "@/shared/schemas/env"
+import z from "zod"
+
+export type EnvData = z.infer<typeof EnvSchema>
+
+export type ConfigData = EnvData
